remove language name use case
